melhorando service e controllers da aplicação

+++ back-end/src/main/java/com/checkInProject/service/evento/EventoService.java
@@ -1,6 +1,9 @@
 package com.checkInProject.service.evento;
 
+import com.checkInProject.dto.EventoDTO;
+import com.checkInProject.model.ETipoUsuario;
 import com.checkInProject.model.Evento;
+import com.checkInProject.model.Usuario;
 import com.checkInProject.repository.evento.EventoRepository;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
@@ -28,26 +31,40 @@ public class EventoService {
     }
 
     @Transactional
-    public Evento criar(Evento evento) {
-        evento.setId(null);
-        return eventoRepository.save(evento);
-    }
+    public Evento criar(Evento evento, Usuario usuario) {
+            if (!(usuario.getRole() == ETipoUsuario.ORGANIZADOR)) {
+                throw new ResponseStatusException(HttpStatus.BAD_REQUEST,"Usuário não é um organizador");
+            }
+            EventoDTO eventoToSave;
+            eventoToSave = EventoDTO.fromEntity(evento);
+            //lembrar de validar se essa lógica faz sentido (do id)
+            eventoToSave.setId(null);
+            return eventoRepository.save(eventoToSave.toEntity());
+        };
 
     @Transactional
-    public Evento atualizar(Long id, Evento dadosAtualizados) {
-        Evento existente = buscarPorId(id);
-        existente.setTitulo(dadosAtualizados.getTitulo());
-        existente.setDescricao(dadosAtualizados.getDescricao());
-        existente.setData(dadosAtualizados.getData());
-        existente.setLocal(dadosAtualizados.getLocal());
-        existente.setVagas(dadosAtualizados.getVagas());
-        existente.setImagemUrl(dadosAtualizados.getImagemUrl());
-        return eventoRepository.save(existente);
+    public Evento atualizar(Long id, Evento dadosAtualizados,  Usuario usuario) {
+        if (!(usuario.getRole() == ETipoUsuario.ORGANIZADOR)) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST,"Usuário não é um organizador");
+
+        }
+        Evento eventoexistente = buscarPorId(id);
+        EventoDTO eventoToUpdate = EventoDTO.fromEntity(eventoexistente);
+        eventoToUpdate.setTitulo(dadosAtualizados.getTitulo());
+        eventoToUpdate.setDescricao(dadosAtualizados.getDescricao());
+        eventoToUpdate.setData(dadosAtualizados.getData());
+        eventoToUpdate.setLocal(dadosAtualizados.getLocal());
+        eventoToUpdate.setVagas(dadosAtualizados.getVagas());
+        eventoToUpdate.setImagemUrl(dadosAtualizados.getImagemUrl());
+        return eventoRepository.save(eventoToUpdate.toEntity());
     }
 
     @Transactional
     public void remover(Long id) {
         Evento existente = buscarPorId(id);
-        eventoRepository.delete(existente);
+        if (existente != null) {
+            eventoRepository.delete(existente);
+        }
+        throw new ResponseStatusException(HttpStatus.NOT_FOUND,"Evento com o ID informado não existe");
     }
 }
